+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased][latest]
+### Fixed
+- Add default reason for unhandled promises (previously reason was "undefined") (#546)
+
 ### Added
 - Add `Honeybadger.clear()` method to clear context and breadcrumbs
 - Deprecate `Honeybadger.resetContext()`