+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased][latest]
+
+## [3.2.0] - 2021-03-30
 ### Fixed
 - Update Node API endpoint to enable source map processing for server-side errors.